+++ FreightAccounting.Core/Model/Remittances/RemittanceReportModel.cs
@@ -7,18 +7,18 @@ public class RemittanceReportModel
     public IEnumerable<RemittanceEntityReportModel> Remittances { get; set; } = new List<RemittanceEntityReportModel>();
 
     //درآمد بین دو تاریخ
-    public int SumIncome { get; set; } 
+    public long SumIncome { get; set; } 
 
     //سود خالص بین دو تاریخ
-    public int SumNetProfit { get; set; } 
+    public long SumNetProfit { get; set; } 
 
     //مالیات پرداختی بین دو تاریخ
-    public int SumTaxPayment { get; set; } 
+    public long SumTaxPayment { get; set; } 
 
     //جمع پرداختی به بیمه
-    public int SumInsurancePayment { get; set; } 
+    public long SumInsurancePayment { get; set; } 
 
     //جمع پورسانت کاربر یا کاربران
-    public int SumUserCut { get; set; } 
+    public long SumUserCut { get; set; } 
 
 }
